test: consolidate TempDir/TempPath into shared test/utils helper
Four download/file-writer unit tests each defined a near-identical local
TempDir or TempPath RAII class, and test_helpers.h carried the
CurrentPid/MakeUniqueTempPath helpers they relied on. Move all of it into a
single test/utils/temp_dir.h in the fl::test namespace and have the tests
use it, removing the duplication Scott flagged on #793.
